requirements.txt
# Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# RAG & orchestration
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
sentence-transformers>=2.2.2
rank_bm25>=0.2.2

# Document loaders
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.11.0

# Config & utils
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0

# UI
streamlit>=1.29.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
